requirements.txt
aiohttp
aiofiles
pathlib2
streamlit
pillow
google-genai
python-dotenv
matplotlib